Inclusao do resumo posicao

+++ kotlin/kotlinEstudos/src/main/resources/db/migration/V1_8__StockUserOperation.sql
@@ -2,7 +2,7 @@ CREATE TABLE STOCK_USER_OPER(
     id VARCHAR(36) NOT NULL PRIMARY KEY,
     usuario_id VARCHAR(36) NOT NULL,
     stock_id VARCHAR(36) NOT NULL,
-    oper_buy_qt NUMERIC(10,0) NOT NULL,
+    oper_buy_qt INTEGER NOT NULL,
     oper_buy_price NUMERIC(10,2) DEFAULT 0,
     oper_buy_date DATE NOT NULL,
     oper_buy_total NUMERIC(10,2) DEFAULT 0,
